ai/requirements.txt
pyyaml
scikit-learn
